backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.65",
    "fastapi>=0.135.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
